requirements-dev.txt
# Development Dependencies for Multi-Tool Chatbot
# Install with: pip install -r requirements-dev.txt

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
httpx>=0.24.0
factory-boy>=3.3.0

# Code Quality & Formatting
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.0
flake8-quotes>=3.3.0
flake8-bugbear>=23.7.0

# Pre-commit Hooks
pre-commit>=3.3.0

# Type Checking
mypy>=1.5.0
types-requests>=2.31.0
types-PyYAML>=6.0.0

# Development Tools
ipython>=8.14.0
ipdb>=0.13.0
jupyter>=1.0.0
notebook>=7.0.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Performance Testing
locust>=2.16.0
pytest-benchmark>=4.0.0

# Security Testing
bandit>=1.7.0
safety>=2.3.0

# Database Testing
pytest-postgresql>=4.1.0
testcontainers>=3.7.0

# Mock & Fixtures
responses>=0.23.0
freezegun>=1.2.0
faker>=19.3.0

# Coverage & Analysis
coverage>=7.2.0
diff-cover>=7.5.0

# Development Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
tqdm>=4.65.0

# API Testing
pytest-fastapi>=0.104.0
fastapi-testclient>=0.1.0

# Async Testing
pytest-asyncio>=0.21.0
aioresponses>=0.7.0

# Environment Management
pip-tools>=7.3.0
pipdeptree>=2.13.0

# Code Generation
cookiecutter>=2.1.0
jinja2>=3.1.0

# Monitoring & Debugging
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.0

# Code Metrics
radon>=6.0.0
mccabe>=0.7.0
xenon>=0.8.0

# Git Hooks
husky>=0.1.0

# Docker Development
docker-compose>=1.29.0

# Database Development
alembic>=1.11.0
psycopg2-binary>=2.9.0

# API Documentation
fastapi-openapi>=0.1.0
fastapi-swagger-ui>=0.1.0

# Development Server
uvicorn[standard]>=0.23.0
watchdog>=3.0.0

# Environment Variables
python-decouple>=3.8.0

# Logging Development
structlog>=23.1.0
colorlog>=6.7.0

# Data Validation
marshmallow>=3.20.0
cerberus>=1.3.0

# Configuration Management
dynaconf>=3.2.0
hydra-core>=1.3.0

# Development Scripts
click>=8.1.0
typer>=0.9.0
rich>=13.4.0

# Code Generation & Templates
cookiecutter>=2.1.0
jinja2>=3.1.0

# Development Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# File Handling
pathlib2>=2.3.0
pathspec>=0.11.0

# Network Utilities
requests-mock>=1.11.0
responses>=0.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization (for development/debugging)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Database
sqlite3  # Built-in with Python
